Reflection:
So the prototype keyword creates a new instance of the superclass (the prototype) and allows you to add new methods onto the class (the prototype.method_name).  The prototype is important because it has methods that the superclass doesn't have access to.

I also learned how to display the object and its properties by console logging the superclass and its prototype to the console.  This allowed me to see what was going on under the hood in regards to the super class and its prototype methods.